Backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Validação e schemas
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Autenticação e segurança
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Processamento de arquivos
pandas==2.1.3
openpyxl==3.1.2

# Matching fuzzy
rapidfuzz==3.5.2

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1

